fix: fixed user part for test_official
- fixed missing check on `UserIsAdminError` in `UserController.deleteUser`
- fixed incorrect check in `DELETE ezelectronics/users/:username` (isAdmin -> isLoggedIn)
- fixed missing `isLoggedIn` check in `PATCH

+++ code/server/src/controllers/userController.ts
@@ -92,6 +92,10 @@ class UserController {
             throw new UserNotAdminError();
         }
 
+        if ((await this.dao.getUserByUsername(username)).role === "Admin") {
+            throw new UserIsAdminError();
+        }
+
         return this.dao.deleteUser(username);
     }
